디렉터리 생성 오류 수정
