+++ SERVER/user.config
@@ -1,0 +1,1 @@
+abc:abc:abc
